requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.1
astroid==2.11.5
async-generator==1.10
asyncpg==0.25.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
coverage==6.3.2
decorator==5.1.1
dill==0.3.4
fastapi==0.77.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
iniconfig==1.1.1
install==1.3.5
iso8601==1.0.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
outcome==1.1.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pylint==2.13.8
pyparsing==3.0.9
pypika-tortoise==0.1.4
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-tornasync==0.6.0.post2
pytest-trio==0.7.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
starlette==0.19.1
tomli==2.0.1
tornado==6.1
tortoise-orm==0.19.0
trio==0.20.0
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
